Allow for uploading of files other than images
Basically it extends the current functionality by checking if the uploaded file is an image, and if it isn't, use the new code.

* The server has to give back an extra value, called "response.document.img".
* When the user has text selected and uploads a document, the selected text is converted to a link
* If there is an image in the selection, that image is wrapped in a link to the document.
* If there is no selection, a new image is used with the src set to response.document.img, on the server side I then check for the mime type and pass back the appropriate image url, for example a small PDF image, or Word image.

This needs a lot of work, and it would be best to extract this out into a separate image uploader and document uploader (because the image previewing feature (during uploading) now shows a broken image).