Rewrite handlePaste method to properly clean pasted content
This solves the problem with diff'ing to clean the pasted content. On a paste event, the cursor is now moved to a hidden element where the paste event will occur. After this, the content is cleaned (using the new whiteListTags option or the "nohtml" option) and the cleaned content is pasted back into the original element. This closes issue #23.