Allow custom dataType on page save
This allows you to set the dataType that is expected to be returned by the server when saving a page.

This can be helpful when you want the server to return some custom javascript action instead of a json value.

You can read more about this at http://api.jquery.com/jQuery.ajax/#jQuery-ajax-settings