Allow the scope of a region to be specified via the 'data-scope' attribute.
Within the context of a site, a content region within a page could belong specifically to the page or globally to the site. A global/site content region would be shared with multiple pages within the site.

For example, given the following content regions on a page:

<div id="region_1" class="mercury-region" data-type="editable" data-scope="site"></div>
<div id="region_2" class="mercury-region" data-type="editable" data-scope="page"></div>

Editing the content in region_1 will affect all pages that include region_1, since it has been denoted as a 'site' region via the 'data-scope' attribute. Editing the content in region_2 will only affect the page that is being edited. The 'site' and 'page' scopes used here are just examples. The host application would define and handle the scopes it supports.